% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_kpi.R
\name{fit_kpi}
\alias{fit_kpi}
\title{Function for determining Kpi from FRRF data}
\usage{
fit_kpi(file_linc, file_control, file_calib, fit_start = 2, fit_finish = 8,
  recovery_start = 8, recovery_finish = 10, light_steps = 10)
}
\arguments{
\item{file_linc}{path to the file with the lincomycin data}

\item{file_control}{path to the file with the control data (without inhibitor)}

\item{file_calib}{path to the calibration file (specific for each machine)}

\item{fit_start}{light step where to start the fit}

\item{fit_finish}{light step where to finish the fit}

\item{recovery_start}{light step where the recovery period starts}

\item{recovery_finish}{light step where the recovery period finishes}

\item{light_steps}{number of light steps inside the file, default=10}
}
\value{
Returns the results as a dataframe and produces a text file with them.
}
\description{
Function for determining Kpi from FRRF data
}
